Add files via upload
Years ago, I searched for a small, simple program to measure my notebook's battery life, but I could not find a good one. So I programmed it myself. A few days ago, I wanted to test again a notebook. I still could not find a program in the internet, so now i decided to do this upload so other people can test their batteries. 

The program is NOT perfect and may, or will contain little bugs. But at all, it seems to work fine.